refactor: split CtResourceCell and CtToggleSwitch catalog parts (#3878)
Extract format/trailing-cluster and state/track/knob library parts so the
largest shared Ct-* widgets stay under file-size targets; document the new
parts in repo-and-packages.md.

Refs #3878
