+++ app/lib/widgets/ct_toggle_switch.dart
@@ -2,6 +2,10 @@ import 'package:flutter/material.dart';
 
 import '../config/editorial_monocle_palette.dart';
 
+part 'ct_toggle_switch_state.dart';
+part 'ct_toggle_switch_track.dart';
+part 'ct_toggle_switch_knob.dart';
+
 /// Two-state pixel-art toggle for the dark editorial-monocle theme.
 ///
 /// Implements `Refs #2859` R8 + § *CtToggleSwitch visual contract*. The
@@ -87,193 +91,4 @@ class CtToggleSwitch extends StatefulWidget {
 
   @override
   State<CtToggleSwitch> createState() => _CtToggleSwitchState();
-}
-
-class _CtToggleSwitchState extends State<CtToggleSwitch> {
-  bool _hovered = false;
-
-  bool get _enabled => widget.onChanged != null;
-
-  void _handleHover(bool entered) {
-    if (!_enabled) return;
-    if (_hovered == entered) return;
-    setState(() => _hovered = entered);
-  }
-
-  void _handleTap() {
-    final ValueChanged<bool>? cb = widget.onChanged;
-    if (cb == null) return;
-    cb(!widget.value);
-  }
-
-  Color get _resolvedGlowColor =>
-      widget.onGlowColor ?? EditorialMonoclePalette.accent;
-
-  @override
-  Widget build(BuildContext context) {
-    if (!_enabled) {
-      return Opacity(
-        opacity: CtToggleSwitch.disabledOpacity,
-        child: _CtToggleSwitchTrack(
-          value: widget.value,
-          hovered: false,
-          animated: false,
-          glowColor: _resolvedGlowColor,
-        ),
-      );
-    }
-    return MouseRegion(
-      onEnter: (_) => _handleHover(true),
-      onExit: (_) => _handleHover(false),
-      cursor: SystemMouseCursors.click,
-      child: GestureDetector(
-        behavior: HitTestBehavior.opaque,
-        onTap: _handleTap,
-        child: _CtToggleSwitchTrack(
-          value: widget.value,
-          hovered: _hovered,
-          animated: true,
-          glowColor: _resolvedGlowColor,
-        ),
-      ),
-    );
-  }
-}
-
-class _CtToggleSwitchTrack extends StatelessWidget {
-  const _CtToggleSwitchTrack({
-    required this.value,
-    required this.hovered,
-    required this.animated,
-    required this.glowColor,
-  });
-
-  final bool value;
-  final bool hovered;
-  final bool animated;
-  final Color glowColor;
-
-  Color get _trackFill => value
-      ? EditorialMonoclePalette.surfaceLite
-      : EditorialMonoclePalette.surface;
-
-  Color get _trackBorder => value
-      ? EditorialMonoclePalette.accent
-      : EditorialMonoclePalette.accentDim;
-
-  Color get _knobFill {
-    if (value) {
-      return hovered
-          ? EditorialMonoclePalette.accentBright
-          : EditorialMonoclePalette.accent;
-    }
-    return hovered
-        ? EditorialMonoclePalette.accentDim
-        : EditorialMonoclePalette.muted;
-  }
-
-  Color get _knobBorder => value
-      ? EditorialMonoclePalette.accentBright
-      : EditorialMonoclePalette.accentDim;
-
-  double get _knobLeft =>
-      value ? CtToggleSwitch.knobOnOffset : CtToggleSwitch.knobOffOffset;
-
-  double get _glowAlpha {
-    if (!value) return 0;
-    return hovered
-        ? CtToggleSwitch.glowHoverAlpha
-        : CtToggleSwitch.glowRestAlpha;
-  }
-
-  @override
-  Widget build(BuildContext context) {
-    final Duration duration = animated
-        ? CtToggleSwitch.animationDuration
-        : Duration.zero;
-    final double knobTopInset =
-        (CtToggleSwitch.trackHeight - CtToggleSwitch.knobSize) / 2;
-    return SizedBox(
-      width: CtToggleSwitch.trackWidth,
-      height: CtToggleSwitch.trackHeight,
-      child: Stack(
-        clipBehavior: Clip.none,
-        children: <Widget>[
-          AnimatedContainer(
-            key: const ValueKey<String>('ctToggleSwitchTrack'),
-            duration: duration,
-            curve: CtToggleSwitch.animationCurve,
-            width: CtToggleSwitch.trackWidth,
-            height: CtToggleSwitch.trackHeight,
-            decoration: BoxDecoration(
-              color: _trackFill,
-              border: Border.all(
-                color: _trackBorder,
-                width: CtToggleSwitch.borderWidth,
-              ),
-            ),
-          ),
-          AnimatedPositioned(
-            duration: duration,
-            curve: CtToggleSwitch.animationCurve,
-            left: _knobLeft,
-            top: knobTopInset,
-            child: _CtToggleSwitchKnob(
-              fill: _knobFill,
-              border: _knobBorder,
-              glowColor: glowColor,
-              glowAlpha: _glowAlpha,
-              duration: duration,
-            ),
-          ),
-        ],
-      ),
-    );
-  }
-}
-
-class _CtToggleSwitchKnob extends StatelessWidget {
-  const _CtToggleSwitchKnob({
-    required this.fill,
-    required this.border,
-    required this.glowColor,
-    required this.glowAlpha,
-    required this.duration,
-  });
-
-  final Color fill;
-  final Color border;
-  final Color glowColor;
-  final double glowAlpha;
-  final Duration duration;
-
-  @override
-  Widget build(BuildContext context) {
-    final List<BoxShadow>? glow = glowAlpha > 0
-        ? <BoxShadow>[
-            BoxShadow(
-              color: glowColor.withValues(
-                alpha: glowAlpha,
-              ),
-              spreadRadius: CtToggleSwitch.glowSpread,
-              blurRadius: 0,
-            ),
-          ]
-        : null;
-    return AnimatedContainer(
-      key: const ValueKey<String>('ctToggleSwitchKnob'),
-      duration: duration,
-      curve: CtToggleSwitch.animationCurve,
-      width: CtToggleSwitch.knobSize,
-      height: CtToggleSwitch.knobSize,
-      decoration: BoxDecoration(
-        color: fill,
-        border: Border.all(
-          color: border,
-          width: CtToggleSwitch.borderWidth,
-        ),
-        boxShadow: glow,
-      ),
-    );
-  }
 }
